Consolidate CharacterRaids and CharacterDungeons into CharacterInstances
