--- frontend/src/components/CharacterRaids.js
@@ -1,40 +1,0 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchInstances } from '../store/instances';
-import ExpansionDropdown from './ExpansionDropdown';
-import '../styles/CharacterInstances.scss'
-
-function CharacterRaids() {
-	const dispatch = useDispatch();
-	const currentCharKey = useSelector((state) => state.session.currentChar);
-	const currentChar = useSelector((state) => state.characters[currentCharKey]);
-	const charRaids = useSelector((state) => state.raids[currentCharKey]);
-	const oAuth = useSelector((state) => state.session.oAuth);
-
-	useEffect(
-		() => {
-			// Prevents refetching of raids if already present in Redux store
-			if (!charRaids) {
-				dispatch(fetchInstances(currentChar.region, currentChar.realm.slug, currentChar.name.toLowerCase(), oAuth, 'raid'));
-			}
-		},
-		[ dispatch, oAuth, currentChar, charRaids ]
-	);
-
-	return (
-		<div className='instance-details border'>
-			<header>Raid Details</header>
-			{charRaids ? (
-				<ul>
-          {Object.entries(charRaids).map(([expansionName, expansion]) => {
-            return <ExpansionDropdown expansion={expansion} expansionName={expansionName} type='raid' key={expansionName} />
-          })}
-				</ul>
-			) : (
-				<p>Fetching Raids...</p>
-			)}
-		</div>
-	);
-}
-
-export default CharacterRaids;
